requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
gradio>=4.0.0
pandas>=2.0.0
numpy>=1.24.0

# ── Visualization ──────────────────────────────────────────────────────────────
plotly>=5.18.0
kaleido>=0.2.1

# ── PDF Generation ─────────────────────────────────────────────────────────────
reportlab>=4.0.0

# ── Data & Tables ──────────────────────────────────────────────────────────────
tabulate>=0.9.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9  # PostgreSQL driver

# ── Authentication ─────────────────────────────────────────────────────────────
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# ── Environment & Config ───────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Google Gemini API ──────────────────────────────────────────────────────────
google-generativeai>=0.3.0

# ── API Integrations (Future) ──────────────────────────────────────────────────
# google-ads>=23.0.0
# facebook-business>=18.0.0

# ── Utilities ──────────────────────────────────────────────────────────────────
requests>=2.31.0
python-dateutil>=2.8.2
